fix(lock): verify remote target lock actually acquired; consolidate integration tests
The target lock never confirmed that the remote `flock -n` acquired: it returned
the started process even when flock exited immediately because the lock was held,
so the orchestrator treated a busy target as "locked" and let a second sync
proceed. Unit tests mocked this path, so it went unnoticed; the new flock-over-SSH
integration test surfaced it against real VMs.

- start_persistent_remote_lock now checks the process exit status after starting
  it: a process that has already exited means flock could not acquire the lock →
  return None (target busy). RemoteProcess gains poll() for this.
- Consolidate the folder-sync scenario into the canonical end-to-end test
  (TestEndToEndSync.test_core_us_job_arch_as1): one seed, five syncs — W1 gate
  block, dry-run rehearsal, real A→B (job integration + full /home metadata,
  ownership, permissions incl. special bits, ACL, mtime, hard/sym links across
  user/root/other-user files and dirs, exclusions, ADR-016 runtime sentinels),
  B→A round-trip (incl. directory-deletion), and a clean A→B. Delete the now-merged
  test_folder_sync.py.
- Lock integration test holds the target lock via a concurrent still-running
  command (keeps the SSH channel open) instead of a detached process that got
  reaped before the sync reached the target lock.
- reset-vm.sh: re-learn the peer VM's current host key on every reset
  (ssh-keyscan) so inter-VM trust self-heals rather than trusting a possibly-stale
  @home baseline — fixes reverse-direction (pc2→pc1) sync failures in CI.
